AQA_lesson/python_api_tests/requirements.txt
# API Testing
requests==2.32.3
pytest==8.3.4
pytest-html==4.1.1
pytest-xdist==3.6.1

# Validation
jsonschema==4.23.0
pydantic==2.10.6

# Reporting
allure-pytest==2.13.5

# Utilities
python-dotenv==1.0.1
faker==33.3.1

# Development
black==24.10.0
flake8==7.1.1

